Update the version after changing the directory naming convention
NOTE: This likely should have been a more significant version update,
but due to this repo still being in the early stages of continuous
development, it is very volatile. That being said, majority of the
version updates will likely be major, so until the repo is at a
stable point, I am going to stick to mild version updates.